Analytics_MCP/server/requirements.txt
fastmcp>=2.12.0
uvicorn>=0.23.2
starlette>=0.27.0
pyyaml>=6.0.1
asyncpg>=0.29.0